// TvServer.WebFrontend/src/apis/TvControlClient.ts
export enum DeviceStatus {
    _0 = 0,
    _1 = 1,
    _2 = 2,
}
// ...
export class SamsungAdditionalDeviceInfo implements ISamsungAdditionalDeviceInfo {
    frameTVSupport?: string | undefined;
    gamePadSupport?: string | undefined;
    imeSyncedSupport?: string | undefined;
    language?: string | undefined;
    os?: string | undefined;
    powerState?: string | undefined;
    tokenAuthSupport?: string | undefined;
    voiceSupport?: string | undefined;
    wallScreenRatio?: string | undefined;
    wallService?: string | undefined;
    countryCode?: string | undefined;
    description?: string | undefined;
    developerIP?: string | undefined;
    developerMode?: string | undefined;
    duid?: string | undefined;
    firmwareVersion?: string | undefined;
    id?: string | undefined;
    ip?: string | undefined;
    model?: string | undefined;
    modelName?: string | undefined;
    name?: string | undefined;
    networkType?: string | undefined;
    resolution?: string | undefined;
    smartHubAgreement?: string | undefined;
    ssid?: string | undefined;
    type?: string | undefined;
    udn?: string | undefined;
    wifiMac?: string | undefined;

    constructor(data?: ISamsungAdditionalDeviceInfo) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.frameTVSupport = _data["frameTVSupport"];
            this.gamePadSupport = _data["gamePadSupport"];
            this.imeSyncedSupport = _data["imeSyncedSupport"];
            this.language = _data["language"];
            this.os = _data["os"];
            this.powerState = _data["powerState"];
            this.tokenAuthSupport = _data["tokenAuthSupport"];
            this.voiceSupport = _data["voiceSupport"];
            this.wallScreenRatio = _data["wallScreenRatio"];
            this.wallService = _data["wallService"];
            this.countryCode = _data["countryCode"];
            this.description = _data["description"];
            this.developerIP = _data["developerIP"];
            this.developerMode = _data["developerMode"];
            this.duid = _data["duid"];
            this.firmwareVersion = _data["firmwareVersion"];
            this.id = _data["id"];
            this.ip = _data["ip"];
            this.model = _data["model"];
            this.modelName = _data["modelName"];
            this.name = _data["name"];
            this.networkType = _data["networkType"];
            this.resolution = _data["resolution"];
            this.smartHubAgreement = _data["smartHubAgreement"];
            this.ssid = _data["ssid"];
            this.type = _data["type"];
            this.udn = _data["udn"];
            this.wifiMac = _data["wifiMac"];
        }
    }

    static fromJS(data: any): SamsungAdditionalDeviceInfo {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungAdditionalDeviceInfo();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["frameTVSupport"] = this.frameTVSupport;
        data["gamePadSupport"] = this.gamePadSupport;
        data["imeSyncedSupport"] = this.imeSyncedSupport;
        data["language"] = this.language;
        data["os"] = this.os;
        data["powerState"] = this.powerState;
        data["tokenAuthSupport"] = this.tokenAuthSupport;
        data["voiceSupport"] = this.voiceSupport;
        data["wallScreenRatio"] = this.wallScreenRatio;
        data["wallService"] = this.wallService;
        data["countryCode"] = this.countryCode;
        data["description"] = this.description;
        data["developerIP"] = this.developerIP;
        data["developerMode"] = this.developerMode;
        data["duid"] = this.duid;
        data["firmwareVersion"] = this.firmwareVersion;
        data["id"] = this.id;
        data["ip"] = this.ip;
        data["model"] = this.model;
        data["modelName"] = this.modelName;
        data["name"] = this.name;
        data["networkType"] = this.networkType;
        data["resolution"] = this.resolution;
        data["smartHubAgreement"] = this.smartHubAgreement;
        data["ssid"] = this.ssid;
        data["type"] = this.type;
        data["udn"] = this.udn;
        data["wifiMac"] = this.wifiMac;
        return data;
    }
}
// ...
export class SamsungAppData implements ISamsungAppData {
    id?: string | undefined;
    appId?: string | undefined;
    appType?: number;
    icon?: string | undefined;
    isLock?: number;
    name?: string | undefined;

    constructor(data?: ISamsungAppData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.appId = _data["appId"];
            this.appType = _data["appType"];
            this.icon = _data["icon"];
            this.isLock = _data["isLock"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): SamsungAppData {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungAppData();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["appId"] = this.appId;
        data["appType"] = this.appType;
        data["icon"] = this.icon;
        data["isLock"] = this.isLock;
        data["name"] = this.name;
        return data;
    }
}

export interface ISamsungAppData {
    id?: string | undefined;
    appId?: string | undefined;
    appType?: number;
    icon?: string | undefined;
    isLock?: number;
    name?: string | undefined;
}
// ...
export enum SamsungKeypress {
    KEY_POWER = "KEY_POWER",
    KEY_HOME = "KEY_HOME",
    KEY_SOURCE = "KEY_SOURCE",
    KEY_GUIDE = "KEY_GUIDE",
    KEY_TOOLS = "KEY_TOOLS",
    KEY_INFO = "KEY_INFO",
    KEY_UP = "KEY_UP",
    KEY_DOWN = "KEY_DOWN",
    KEY_LEFT = "KEY_LEFT",
    KEY_RIGHT = "KEY_RIGHT",
    KEY_ENTER = "KEY_ENTER",
    KEY_RETURN = "KEY_RETURN",
    KEY_CH_LIST = "KEY_CH_LIST",
    KEY_CHUP = "KEY_CHUP",
    KEY_CHDOWN = "KEY_CHDOWN",
    KEY_VOLUP = "KEY_VOLUP",
    KEY_VOLDOWN = "KEY_VOLDOWN",
    KEY_MUTE = "KEY_MUTE",
    KEY_EXIT = "KEY_EXIT",
    KEY_RED = "KEY_RED",
    KEY_GREEN = "KEY_GREEN",
    KEY_YELLOW = "KEY_YELLOW",
    KEY_BLUE = "KEY_BLUE",
    KEY_FACTORY = "KEY_FACTORY",
}
// ...
export enum SamsungKeypressType {
    Click = "Click",
    Press = "Press",
    Release = "Release",
}
// ...
export class SamsungTvDto implements ISamsungTvDto {
    id!: string;
    deviceName!: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    samsungApps?: SamsungAppData[] | undefined;
    deviceInfo?: SamsungTvInfo;

    constructor(data?: ISamsungTvDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.deviceName = _data["deviceName"];
            this.lastKnownIp = _data["lastKnownIp"];
            this.status = _data["status"];
            if (Array.isArray(_data["samsungApps"])) {
                this.samsungApps = [] as any;
                for (let item of _data["samsungApps"])
                    this.samsungApps!.push(SamsungAppData.fromJS(item));
            }
            this.deviceInfo = _data["deviceInfo"] ? SamsungTvInfo.fromJS(_data["deviceInfo"]) : undefined as any;
        }
    }

    static fromJS(data: any): SamsungTvDto {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungTvDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["deviceName"] = this.deviceName;
        data["lastKnownIp"] = this.lastKnownIp;
        data["status"] = this.status;
        if (Array.isArray(this.samsungApps)) {
            data["samsungApps"] = [];
            for (let item of this.samsungApps)
                data["samsungApps"].push(item ? item.toJSON() : undefined as any);
        }
        data["deviceInfo"] = this.deviceInfo ? this.deviceInfo.toJSON() : undefined as any;
        return data;
    }
}

export interface ISamsungTvDto {
    id: string;
    deviceName: string;
    lastKnownIp?: string | undefined;
    status?: DeviceStatus;
    samsungApps?: SamsungAppData[] | undefined;
    deviceInfo?: SamsungTvInfo;
}

export class SamsungTvInfo implements ISamsungTvInfo {
    device?: SamsungAdditionalDeviceInfo;
    id?: string | undefined;
    isSupport?: string | undefined;
    name?: string | undefined;
    remote?: string | undefined;
    type?: string | undefined;
    uri?: string | undefined;
    version?: string | undefined;

    constructor(data?: ISamsungTvInfo) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.device = _data["device"] ? SamsungAdditionalDeviceInfo.fromJS(_data["device"]) : undefined as any;
            this.id = _data["id"];
            this.isSupport = _data["isSupport"];
            this.name = _data["name"];
            this.remote = _data["remote"];
            this.type = _data["type"];
            this.uri = _data["uri"];
            this.version = _data["version"];
        }
    }

    static fromJS(data: any): SamsungTvInfo {
        data = typeof data === 'object' ? data : {};
        let result = new SamsungTvInfo();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["device"] = this.device ? this.device.toJSON() : undefined as any;
        data["id"] = this.id;
        data["isSupport"] = this.isSupport;
        data["name"] = this.name;
        data["remote"] = this.remote;
        data["type"] = this.type;
        data["uri"] = this.uri;
        data["version"] = this.version;
        return data;
    }
}

export interface ISamsungTvInfo {
    device?: SamsungAdditionalDeviceInfo;
    id?: string | undefined;
    isSupport?: string | undefined;
    name?: string | undefined;
    remote?: string | undefined;
    type?: string | undefined;
    uri?: string | undefined;
    version?: string | undefined;
}